Organize project structure and add Socket.IO server
