+++ .gitignore
@@ -1,5 +1,6 @@
-//// .gitignorej
 node_modules/
 .env
 client/node_modules/
 client/build/
+server/node_modules/
+
