--- App.js
@@ -1,12 +1,0 @@
-//// client/src/App.js
-import Chat from './components/Chat';
-
-function App() {
-  return (
-    <div className="App">
-      <Chat />
-    </div>
-  );
-}
-
-export default App;
