--- index.html
@@ -1,12 +1,0 @@
-//// client/public/index.html
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>QuickChat</title>
-</head>
-<body>
-  <div id="root"></div>
-</body>
-</html>
