+++ server/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "quickchat-server",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "express": "^4.18.2",
+    "socket.io": "^4.6.0"
+  },
+  "scripts": {
+    "start": "node index.js"
+  }
+}
